Adding some fixes + more tests
